pyproject.toml
[project]
name = "snailly-backend-thesis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "nltk>=3.9.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "sastrawi>=1.0.1",
    "scikit-learn>=1.7.1",
    "selenium>=4.34.2",
    "torch>=2.8.0",
    "transformers>=4.55.0",
    "webdriver-manager>=4.0.2",
]
